+++ Assets/CommonCore/CoreShared/SharedUtils.cs
@@ -33,7 +33,7 @@ namespace CommonCore
             MetaState.Instance.NextScene = sceneOverride;
             MetaState.Instance.TransitionType = SceneTransitionType.NewGame;
             //GC.Collect();            
-            SceneManager.LoadScene(CoreParams.LoadingScene);            
+            CoreUtils.LoadScene(CoreParams.LoadingScene);            
         }
 
         /// <summary>
@@ -42,7 +42,7 @@ namespace CommonCore
         public static void ShowGameOver()
         {
             MetaState.Instance.NextScene = SceneManager.GetActiveScene().name; //in case we need it...
-            SceneManager.LoadScene(CoreParams.GameOverScene);
+            CoreUtils.LoadScene(CoreParams.GameOverScene);
         }
 
         /// <summary>
@@ -63,7 +63,7 @@ namespace CommonCore
             MetaState.Instance.TransitionType = SceneTransitionType.EndGame;
             //GC.Collect();
             TryGetSceneController().Ref()?.ExitScene();
-            SceneManager.LoadScene(CoreParams.LoadingScene);
+            CoreUtils.LoadScene(CoreParams.LoadingScene);
         }
 
         /// <summary>