+++ CHANGELOG.md
@@ -604,3 +604,4 @@
 * Player prefab name can now be overridden from WorldSceneController or GameState
 * Added PersistInitialPosition option to ActorController to save InitialPosition
 * Resource manifest generation now includes sub-assets
+* Add scene override redirection APIs to CoreUtils